ADDED vents, secret walls, liquor puddles, cheat path extension.
The vents do not have teleportation in yet and are not timed yet. The secret walls are destroyed on activator collision with the head and can be duplicated and placed wherever needed. Liquor puddles harm the player, the particle effect does not, and the bottles still have colliders on them and do not harm the player. The cheat path has been extended to the first half of the maze. SCRIPTS MAY BE MESSY AND ALL OVER THE PLACE SORRYYY.