BIG CHANGES
Added text for button activation and door opening with a 5 second timer. Added vent teleportation. Adjusted the floor liners and edges for the bottom of the walls. Planning on placing models in the second half of the maze and the cheat path tomorrow. Saturday is VENT day where i redo all the vents and test them. Playtesting and fixing things as well.