MAZE CHANGES
Added Dareen's yellow spill models only. I have to attach the "hazards" script to them and then add the red spill to the mop variants. Planning on doing something about the fuzzy particle effects cause they look wonkey and blend in too much with the activators for the vents/mops.